✨ feat: Detail 페이지 컴포넌트 세분화 작업
